Move PSA information and dependency automation into their own module
This will let us use these features from other modules (yet to be created).
